+++ src/contracts/migrations/MigrateFloorToken.sol
@@ -2,11 +2,10 @@
 
 pragma solidity ^0.8.0;
 
-import '@openzeppelin/contracts/token/ERC20/ERC20.sol';
+import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
-import '../tokens/Floor.sol';
-import '../../interfaces/migrations/MigrateFloorToken.sol';
-
+import "../tokens/Floor.sol";
+import "../../interfaces/migrations/MigrateFloorToken.sol";
 
 /**
  * Burns FLOOR v1 tokens for FLOOR v2 tokens. We have a list of the defined
@@ -19,30 +18,29 @@ import '../../interfaces/migrations/MigrateFloorToken.sol';
  * should be made prior to calling this contract function.
  */
 contract MigrateFloorToken is IMigrateFloorToken {
-
     address[] private MIGRATED_TOKENS = [
-        0xf59257E961883636290411c11ec5Ae622d19455e,  // Floor
-        0x0C3983165E9BcE0a9Bb43184CC4eEBb26dce48fA,  // aFloor
-        0xb1Cc59Fc717b8D4783D41F952725177298B5619d,  // gFloor
-        0x164AFe96912099543BC2c48bb9358a095Db8e784   // sFloor
+        0xf59257E961883636290411c11ec5Ae622d19455e, // Floor
+        0x0C3983165E9BcE0a9Bb43184CC4eEBb26dce48fA, // aFloor
+        0xb1Cc59Fc717b8D4783D41F952725177298B5619d, // gFloor
+        0x164AFe96912099543BC2c48bb9358a095Db8e784 // sFloor
     ];
 
-    event FloorMigrated(address caller, uint amount);
+    event FloorMigrated(address caller, uint256 amount);
 
     address public immutable newFloor;
 
-    constructor (address _newFloor) {
+    constructor(address _newFloor) {
         newFloor = _newFloor;
     }
 
-    function mintTokens(uint _amount) external override {
+    function mintTokens(uint256 _amount) external override {
         FLOOR(newFloor).mint(address(this), _amount);
     }
 
     function upgradeFloorToken() external override {
         // Keep a running total of allocated tokens
-        uint floorAllocation;
-        uint tokenBalance;
+        uint256 floorAllocation;
+        uint256 tokenBalance;
         IERC20 token;
 
         // Loop through the tokens
@@ -64,14 +62,15 @@ contract MigrateFloorToken is IMigrateFloorToken {
                 floorAllocation += tokenBalance;
             }
 
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
 
-        require(floorAllocation > 0, 'No tokens available to migrate');
+        require(floorAllocation > 0, "No tokens available to migrate");
 
         // Mint our FLOOR tokens to the sender
         FLOOR(newFloor).mint(msg.sender, floorAllocation);
         emit FloorMigrated(msg.sender, floorAllocation);
     }
-
 }
