+++ src/contracts/options/Option.sol
@@ -4,20 +4,19 @@ pragma solidity ^0.8.0;
 
 import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
 import "@openzeppelin/contracts/utils/Counters.sol";
-import '@openzeppelin/contracts/utils/Base64.sol';
-import '@openzeppelin/contracts/utils/Counters.sol';
+import "@openzeppelin/contracts/utils/Base64.sol";
+import "@openzeppelin/contracts/utils/Counters.sol";
 
-import '../../interfaces/options/Option.sol';
+import "../../interfaces/options/Option.sol";
 
 contract Option is ERC721URIStorage {
-
     using Counters for Counters.Counter;
 
     // We cannot just use balanceOf to create the new tokenId because tokens
     // can be burned (destroyed), so we need a separate counter.
     Counters.Counter private _tokenIdTracker;
 
-    constructor() ERC721('FloorOption', 'FOPT') {}
+    constructor() ERC721("FloorOption", "FOPT") {}
 
     /**
      * The DNA of our Option defines the struct of our Option, but without the
@@ -40,40 +39,40 @@ contract Option is ERC721URIStorage {
      *  return (_c << 4) | _d;
      * }
      */
-    mapping (uint => bytes32) private dna;
+    mapping(uint256 => bytes32) private dna;
 
     /**
      * Gets the allocation granted to the Option.
      */
-    function allocation(uint tokenId) public view returns (uint) {
+    function allocation(uint256 tokenId) public view returns (uint256) {
         return sliceUint(dna[tokenId], 0);
     }
 
     /**
      * Gets the reward amount granted to the Option.
      */
-    function rewardAmount(uint tokenId) public view returns (uint) {
+    function rewardAmount(uint256 tokenId) public view returns (uint256) {
         return sliceUint(dna[tokenId], 8);
     }
 
     /**
      * Gets the rarity of the Option, calculated at point of mint.
      */
-    function rarity(uint tokenId) public view returns (uint) {
+    function rarity(uint256 tokenId) public view returns (uint256) {
         return sliceUint(dna[tokenId], 16);
     }
 
     /**
      * Gets the pool ID that the Option is attributed to.
      */
-    function poolId(uint tokenId) public view returns (uint) {
+    function poolId(uint256 tokenId) public view returns (uint256) {
         return sliceUint(dna[tokenId], 20);
     }
 
     /**
      * Takes a bytes input and converts it to an integer
      */
-    function sliceUint(bytes32 bs, uint start) internal pure returns (uint x) {
+    function sliceUint(bytes32 bs, uint256 start) internal pure returns (uint256 x) {
         assembly {
             x := mload(add(bs, add(0x20, start)))
         }
@@ -94,14 +93,11 @@ contract Option is ERC721URIStorage {
      */
     function baseURI() public pure returns (string memory) {}
 
-
-    function _beforeTokenTransfer(
-        address from,
-        address to,
-        uint256 firstTokenId,
-        uint256 batchSize
-    ) internal virtual override {
+    function _beforeTokenTransfer(address from, address to, uint256 firstTokenId, uint256 batchSize)
+        internal
+        virtual
+        override
+    {
         super._beforeTokenTransfer(from, to, firstTokenId, batchSize);
     }
-
 }
