+++ src/interfaces/actions/Action.sol
@@ -2,9 +2,6 @@
 
 pragma solidity ^0.8.0;
 
-
 interface IAction {
-
-    function execute(bytes calldata) external returns (uint);
-
+    function execute(bytes calldata) external returns (uint256);
 }
